+++ src/main/java/com/ahex/match/pojo/Run.java
@@ -3,34 +3,32 @@ package com.ahex.match.pojo;
 
 public class Run{
 		
-	private Integer batsman;
-	private Integer extras;
-	private Integer total;
-	private Integer non_boundary;
-	
-	public Integer getBatsman() {
+	private Long batsman;
+	private Long extras;
+	private Long total;
+	private Long non_boundary;
+	public Long getBatsman() {
 		return batsman;
 	}
-	public void setBatsman(Integer batsman) {
+	public void setBatsman(Long batsman) {
 		this.batsman = batsman;
 	}
-	public Integer getExtras() {
+	public Long getExtras() {
 		return extras;
 	}
-	public void setExtras(Integer extras) {
+	public void setExtras(Long extras) {
 		this.extras = extras;
 	}
-	public Integer getTotal() {
+	public Long getTotal() {
 		return total;
 	}
-	public void setTotal(Integer total) {
+	public void setTotal(Long total) {
 		this.total = total;
 	}
-	
-	public Integer getNon_boundary() {
+	public Long getNon_boundary() {
 		return non_boundary;
 	}
-	public void setNon_boundary(Integer non_boundary) {
+	public void setNon_boundary(Long non_boundary) {
 		this.non_boundary = non_boundary;
 	}
 	@Override
@@ -38,6 +36,8 @@ public class Run{
 		return "Run [batsman=" + batsman + ", extras=" + extras + ", total=" + total + ", non_boundary=" + non_boundary
 				+ "]";
 	}
+	
+	
 	
 	
 	}
